Move keep-alive thread to EventSourceBeaconChainEventAdapter since it's only needed when that's in use.
